+++ dca/tl_module.php
@@ -11,13 +11,12 @@
  * @license LGPL
  */
 
-
 /**
  * Add palettes to tl_module
  */
 $GLOBALS['TL_DCA']['tl_module']['palettes']['eventsubscribe']     = '{title_legend},name,headline,type;{redirect_legend:hide},jumpTo_subscribe,jumpTo_unsubscribe,jumpTo_login;{protected_legend:hide},protected;{expert_legend:hide},guests,cssID,space';
 $GLOBALS['TL_DCA']['tl_module']['palettes']['eventlistsubscribe'] = '{title_legend},name,headline,type;{config_legend},cal_calendar,cal_order,cal_limit,perPage;{redirect_legend:hide},jumpTo_subscribe,jumpTo_unsubscribe,jumpTo_login;{template_legend:hide},cal_template,imgSize;{protected_legend:hide},protected;{expert_legend:hide},guests,cssID,space';
-
+$GLOBALS['TL_DCA']['tl_module']['palettes']['eventsubscriptions'] = '{title_legend},name,headline,type;{config_legend},cal_calendar,cal_noSpan,cal_format,cal_ignoreDynamic,cal_order,cal_readerModule,cal_limit,perPage;{template_legend:hide},cal_template,customTpl;{image_legend:hide},imgSize;{protected_legend:hide},protected;{expert_legend:hide},guests,cssID,space';
 
 /**
  * Add fields to tl_module