+++ languages/en/modules.php
@@ -11,9 +11,15 @@
  * @license LGPL
  */
 
-
 /**
  * Front end modules
  */
-$GLOBALS['TL_LANG']['FMD']['eventsubscribe']     = array('Subscribe event form', 'Adds a subscribe event form to the page.');
-$GLOBALS['TL_LANG']['FMD']['eventlistsubscribe'] = array('Event list subscribe', 'Adds a list of subscribable events to the page.');
+$GLOBALS['TL_LANG']['FMD']['eventsubscribe']     = ['Subscribe event form', 'Adds a subscribe event form to the page.'];
+$GLOBALS['TL_LANG']['FMD']['eventlistsubscribe'] = [
+    'Event list subscribe',
+    'Adds a list of subscribable events to the page.',
+];
+$GLOBALS['TL_LANG']['FMD']['eventsubscriptions'] = [
+    'User event subscriptions',
+    'Adds a list of events that the current user is subscribed to.',
+];